Add jsdoc to AnimationStatus and AnimationHandle
