+++ scripts/snapshots/es6Files.txt
@@ -12,6 +12,7 @@
   "es6/animation/easing.js",
   "es6/animation/matchBy.js",
   "es6/animation/timeoutController.js",
+  "es6/animation/types.js",
   "es6/animation/useAnimationManager.js",
   "es6/animation/useAnimationStartSnapshot.js",
   "es6/animation/util.js",
